# MinecraftMapper/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Minecraft Mapper API",
    "version": "v1"
  },
  "paths": {
    "/api/LocationTypes": {
      "get": {
        "tags": [
          "LocationTypes"
        ],
        "operationId": "GetLocationTypes",
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Map/{maximumRingNumber}": {
      "get": {
        "tags": [
          "Map"
        ],
        "operationId": "GetMaps",
        "parameters": [
          {
            "name": "maximumRingNumber",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Realm": {
      "get": {
        "tags": [
          "Realm"
        ],
        "operationId": "GetRealms",
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Realm"
        ],
        "operationId": "AddRealm",
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/NewRealmRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewRealmRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/NewRealmRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/NewRealmRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Realm/{realmId}": {
      "get": {
        "tags": [
          "Realm"
        ],
        "operationId": "GetRealm",
        "parameters": [
          {
            "name": "realmId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Realm"
        ],
        "operationId": "DeleteRealm",
        "parameters": [
          {
            "name": "realmId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Realm/{realmId}/location": {
      "post": {
        "tags": [
          "Realm"
        ],
        "operationId": "AddLocation",
        "parameters": [
          {
            "name": "realmId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/NewLocationRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewLocationRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/NewLocationRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/NewLocationRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Realm/{realmId}/location/{locationId}": {
      "delete": {
        "tags": [
          "Realm"
        ],
        "operationId": "DeleteLocation",
        "parameters": [
          {
            "name": "realmId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "locationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Realm"
        ],
        "operationId": "UpdateLocation",
        "parameters": [
          {
            "name": "realmId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "locationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLocationRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLocationRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLocationRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLocationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "NewRealmRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "NewLocationRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "x": {
            "type": "integer",
            "format": "int32"
          },
          "z": {
            "type": "integer",
            "format": "int32"
          },
          "y": {
            "type": "integer",
            "format": "int32"
          },
          "locationTypeId": {
            "type": "string",
            "format": "uuid"
          },
          "hasPortal": {
            "type": "boolean"
          },
          "hasEnderChest": {
            "type": "boolean"
          },
          "hasEnchantmentTable": {
            "type": "boolean"
          },
          "hasBed": {
            "type": "boolean"
          },
          "hasAnvil": {
            "type": "boolean"
          },
          "hasFurnace": {
            "type": "boolean"
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UpdateLocationRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "locationTypeId": {
            "type": "string",
            "format": "uuid"
          },
          "hasPortal": {
            "type": "boolean"
          },
          "hasEnderChest": {
            "type": "boolean"
          },
          "hasEnchantmentTable": {
            "type": "boolean"
          },
          "hasBed": {
            "type": "boolean"
          },
          "hasAnvil": {
            "type": "boolean"
          },
          "hasFurnace": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      }
    }
  }
}